+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - give dabloon rewards for using global boosters
 - add evil larry
 - prevent duplicate broken rings from simultaneous divorces
+- improved reliability of market system and moved $mk create to $mk buy|sell <item> <amount> <price/item>
 
 # july 2026